requirements.txt
Django==3.2.5
psycopg2-binary==2.9.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
requests==2.31.0
urllib3==1.26.12